server/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.2
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.3.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
Markdown==3.7
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
